# docs/api-specs/delivery-api.openapi.yaml
openapi: 3.0.3
info:
  title: CarGoro Delivery API
  description: |
    차량 탁송 관리, 기사 배정, 루트 관리 API
  version: 0.1.0
  contact:
    name: CarGoro 개발팀
    email: dev@cargoro.com
servers:
  - url: 'https://api.cargoro.com/delivery'
    description: 운영 환경
  - url: 'https://api-staging.cargoro.com/delivery'
    description: 스테이징 환경
  - url: 'http://localhost:8004'
    description: 로컬 개발 환경
...
paths:
...
  /drivers/schedules:
    get:
      summary: 기사 일정 목록 조회
      description: 기사 일정 목록을 조회합니다.
      tags:
        - drivers
      parameters:
        - $ref: '#/components/parameters/DriverID'
        - name: date
          in: query
          description: 조회할 날짜 (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: is_available
          in: query
          description: 가용 상태 필터
          schema:
            type: boolean
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 기사 일정 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DriverScheduleResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: 기사 일정 생성
      description: 새로운 기사 일정을 생성합니다.
      tags:
        - drivers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverScheduleCreate'
      responses:
        '201':
          description: 생성된 기사 일정
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverScheduleResponse'
        '400':
          description: 잘못된 요청 (기사 ID 등 필수 필드 누락)
        '500':
          $ref: '#/components/responses/InternalServerError'
  /drivers/schedules/{schedule_id}:
    get:
      summary: 기사 일정 조회
      description: ID로 특정 기사 일정을 조회합니다.
      tags:
        - drivers
      parameters:
        - $ref: '#/components/parameters/ScheduleID'
      responses:
        '200':
          description: 기사 일정 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverScheduleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      summary: 기사 일정 수정
      description: 특정 기사 일정의 정보를 수정합니다.
      tags:
        - drivers
      parameters:
        - $ref: '#/components/parameters/ScheduleID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverScheduleUpdate'
      responses:
        '200':
          description: 수정된 기사 일정
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverScheduleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: 기사 일정 삭제
      description: 특정 기사 일정을 삭제합니다.
      tags:
        - drivers
      parameters:
        - $ref: '#/components/parameters/ScheduleID'
      responses:
        '200':
          description: 삭제된 기사 일정
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverScheduleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
...
    DriverID:
      name: driver_id
      in: query
      description: 기사 ID
      schema:
        type: string
        format: uuid
...
    ScheduleID:
      name: schedule_id
      in: path
      description: 기사 일정 ID
      required: true
      schema:
        type: string
        format: uuid
...
    Skip:
      name: skip
      in: query
      description: 건너뛸 항목 수
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      description: 반환할 최대 항목 수
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 100
  responses:
    NotFound:
      description: 요청한 리소스를 찾을 수 없습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    DriverScheduleCreate:
      type: object
      required:
        - driver_id
        - date
        - start_time
        - end_time
      properties:
        driver_id:
          type: string
          format: uuid
          description: 기사 ID
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        date:
          type: string
          format: date
          description: 날짜
          example: "2023-05-25"
        start_time:
          type: string
          format: time
          description: 시작 시간
          example: "09:00:00"
        end_time:
          type: string
          format: time
          description: 종료 시간
          example: "18:00:00"
        is_available:
          type: boolean
          description: 가용 여부
          example: true
        reason:
          type: string
          description: 사유 (비가용 시)
          example: "휴가"
        notes:
          type: string
          description: 메모
          example: "오전만 가능"
    DriverScheduleUpdate:
      type: object
      properties:
        date:
          type: string
          format: date
          description: 날짜
        start_time:
          type: string
          format: time
          description: 시작 시간
        end_time:
          type: string
          format: time
          description: 종료 시간
        is_available:
          type: boolean
          description: 가용 여부
        reason:
          type: string
          description: 사유 (비가용 시)
        notes:
          type: string
          description: 메모
    DriverScheduleResponse:
      allOf:
        - $ref: '#/components/schemas/DriverScheduleCreate'
        - type: object
          required:
            - id
            - created_at
            - updated_at
          properties:
            id:
              type: string
              format: uuid
              description: 일정 ID
              example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
            created_at:
              type: string
              format: date-time
              description: 생성 일시
              example: "2023-05-20T12:34:56Z"
            updated_at:
              type: string
              format: date-time
              description: 수정 일시
              example: "2023-05-20T12:34:56Z"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: HTTP 상태 코드
          example: 404
        message:
          type: string
          description: 오류 메시지
          example: "ID 'f47ac10b-58cc-4372-a567-0e02b2c3d479'인 탁송을 찾을 수 없습니다."
        details:
          type: string
          description: 추가 오류 상세 정보
          example: "데이터베이스에서 해당 ID에 해당하는 레코드가 없습니다." 